Move `mephenger/config/config.py` -> `mephenger/config.py`
