+++ mephenger/config.py
@@ -5,7 +5,6 @@
     Ce fichier contient toutes les variables communes au projet
 """
 
-# Couleurs
 import os
 
 MAIN_COLOR = (0.580, 0.533, 0.984)
@@ -13,7 +12,6 @@ R, G, B, = MAIN_COLOR
 BG_COLOR_LEVEL_1 = (R, G, B, 1)
 TEXT_COLOR = (1, 1, 1, 1)
 
-# Chemins
 ROOT_DIR = os.environ["PRJ_ROOT"]
 PUBLIC_DIR = os.path.join(ROOT_DIR, 'public')
 IMG_DIR = os.path.join(PUBLIC_DIR, 'images')
